consolidate unauthed e2e tests (#6668)
